==> 02_xanes_analysis.qmd <==
---
title: "02_xanes_analysis"
format:
  html:
    fig-height: 7
    fig-width: 7
editor: visual
editor_options: 
  chunk_output_type: console
---

## Setup

```{r}
library(tidyverse)
library(gt)
source("source/read_mu_file.R")
source("source/read_many_mu_files.R")
source("source/theme_xas.R")
```

```{r}
xanes <- read_rds("cache/xanes_data_raw.RDS")

metadata <- readxl::read_excel("data/metadata_annotated.xlsx", na = "NA")
```

Join xanes data and metadata

```{r}
xanes_w_metadata <- xanes |> 
  left_join(metadata, by = join_by(filename))

#saveRDS(xanes_w_metadata, file = "cache/xanes_w_metadata.rds")
#writexl::write_xlsx(xanes_w_metadata, path = "cache/xanes_w_metadata.xlsx")
```

# Find E0 positions

Here we define E0 as the first point in the normalized spectrum that reaches 0.5 of the post-edge norm value `1`. To find E0s, we will conduct a linear interpolation of our XANES spectra to provide a more fine-grained assessment of E0 location.

```{r}
e0s <- xanes_w_metadata |> 
  select(filename, energy, samnorm) |> # select down to columns we want
  group_by(filename) |> # for each unique file ...
  # reframing the dataset changes the data to an arbitrary number of columns
  # and returns an ungrouped data frame
  reframe(
    energy_interp = approx(x = energy, y = samnorm, n = 10000)$x,
    samnorm_interp = approx(x = energy, y = samnorm, n = 10000)$y
  ) |> 
  group_by(filename) |> # regroup the data
  arrange(energy_interp) |> # make sure the data is sorted by energy
  filter(samnorm_interp >= 0.5) |> # filter for points above 0.5 norm
  slice(1) |> # find the first point above 0.5 norm. This is our E0
  ungroup() |> # ungroup the data
  rename(e0 = energy_interp, e0_abs = samnorm_interp) # give the e0 values specific names
  
xanes_w_e0s <- xanes_w_metadata |> 
  left_join(e0s, by = join_by(filename))
```

We also want to calculate the peak positions of the k-edge white lines. We do this by finding the first local maximum of the data.

```{r}

```

## TLE

```{r}
xanes_TLE <- xanes_w_e0s |> 
  filter(is_TLE) |> 
  mutate(
    depth_str = paste(core_depth_m, "m"),
    depth_fct = as.factor(depth_str),
    depth_fct = fct_reorder(depth_fct, core_depth_m, .na_rm = FALSE)
  ) |> 
  arrange(depth_fct)
```

### Find peak position of TLE standards

```{r}
xanes_TLE_STD_max <- xanes_TLE |> 
  filter(is_standard) |> 
  group_by(compound) |> 
  # find the eV that is max samnorm
  summarise(
    energy = energy[which.max(samnorm)],
    samnorm = max(samnorm)
    ) |> 
  ungroup()
```

### Plot TLE XANES

```{r}

p_tle_standards <- xanes_TLE |> 
  filter(compound != "BA1B_280") |> 
  filter(is_standard) |> 
  mutate(header = "Standards") |> 
  ggplot(
    aes(
      x = energy,
      y = samnorm,
      color = compound,
      fill = compound
    )
  ) +
  scale_fill_brewer(palette = "Set1") +
  scale_color_brewer(palette = "Set1") +
  geom_vline(
    data = xanes_TLE_STD_max,
    aes(xintercept = energy,
        color = compound),
    linewidth = 0.25,
    show.legend = FALSE
  ) +
  geom_line(linewidth = 1) +
  coord_cartesian(xlim = c(2465, 2490)) +
  ggprism::annotation_ticks(sides = "tb") +
  facet_wrap(vars(header)) +
  labs(
    x = "Energy (eV)",
    y = "Normalized Absorption",
    color = "Standard Compound"
  ) +
  theme_xas() +
  theme()
p_tle_standards

cowplot::save_plot(
  plot = p_tle_standards,
  filename = "fig_output/tle_standards.pdf",
  base_height = 6,
  base_width = 9
)

cowplot::save_plot(
  plot = p_tle_standards,
  filename = "fig_output/tle_standards.png",
  base_height = 6,
  base_width = 9
)
```

Plot TLE Samples

```{r}
p_ba1b <- xanes_TLE |> 
  filter(!is_standard) |> 
  filter(!is_TLE_blank) |> 
  filter(!is_GFF) |> 
  filter(core == "BA1B") |> 
  ggplot(
    aes(
      x = energy,
      y = samnorm,
      color = depth_fct
    )
  ) +
  geom_vline(
    data = xanes_TLE_STD_max,
    aes(xintercept = energy),
    color = "gray",
    linewidth = 0.25
  ) +
  geom_line(linewidth = 1) +
  coord_cartesian(xlim = c(2465, 2490)) +
  scale_color_viridis_d(direction = -1, end = 0.9) +
  facet_wrap(vars(core)) +
  ggprism::annotation_ticks(sides = "tb") +
  labs(
    x = "Energy (eV)",
    y = "Normalized Absorption",
    color = "Core Depth"
  ) +
  theme_xas() +
  theme()
p_ba1b
```

```{r}
p_ba4a <- xanes_TLE |> 
  filter(!is_standard) |> 
  filter(!is_TLE_blank) |> 
  filter(!is_GFF) |> 
  filter(core == "BA4A") |> 
  ggplot(
    aes(
      x = energy,
      y = samnorm,
      color = depth_fct
    )
  ) +
  geom_vline(
    data = xanes_TLE_STD_max,
    aes(xintercept = energy),
    color = "gray",
    linewidth = 0.25
  ) +
  geom_line(linewidth = 1) +
  coord_cartesian(xlim = c(2465, 2490)) +
  scale_color_viridis_d(direction = -1, end = 0.9) +
  facet_wrap(vars(core)) +
  ggprism::annotation_ticks(sides = "tb") +
  labs(
    x = "Energy (eV)",
    y = "Normalized Absorption",
    color = "",
  ) +
  theme_xas() +
  theme()
p_ba4a
```

```{r}
p_ba3a <- xanes_TLE |> 
  filter(!is_standard) |> 
  filter(!is_TLE_blank) |> 
  filter(!is_GFF) |> 
  filter(core == "BA3A") |> 
  ggplot(
    aes(
      x = energy,
      y = samnorm,
      color = depth_fct
    )
  ) +
  geom_vline(
    data = xanes_TLE_STD_max,
    aes(xintercept = energy),
    color = "gray",
    linewidth = 0.25
  ) +
  geom_line(linewidth = 1) +
  coord_cartesian(xlim = c(2465, 2490)) +
  scale_color_viridis_d(direction = -1, end = 0.9) +
  facet_wrap(vars(core)) +
  ggprism::annotation_ticks(sides = "tb") +
  labs(
    x = "Energy (eV)",
    y = "Normalized Absorption",
    color = ""
  ) +
  theme_xas() +
  theme(
    panel.background = element_rect(color = "black"),
    strip.background = element_rect(color = "#464747", fill = "#464747"),
    strip.text = element_text(color = "white"),
    axis.ticks.x = element_blank()
  )
p_ba3a
```

```{r}
p_tle_blanks <- xanes_TLE |> 
  filter(is_TLE_blank) |> 
  mutate(header = "Controls") |> 
  ggplot(
    aes(
      x = energy,
      y = samnorm,
      color = compound
    )
  ) +
  geom_vline(
    data = xanes_TLE_STD_max,
    aes(xintercept = energy),
    linewidth = 0.25,
    color = "gray",
    show.legend = FALSE
  ) +
  geom_line(linewidth = 1) +
  coord_cartesian(xlim = c(2465, 2490)) +
  ggprism::annotation_ticks(sides = "tb") +
  facet_wrap(vars(header)) +
  labs(
    x = "Energy (eV)",
    y = "Normalized Absorption",
    color = "Control"
  ) +
  theme_xas() +
  theme()
p_tle_blanks
```

Export to pdf and png

```{r}
p_tle_combined <- cowplot::plot_grid(p_ba1b, p_ba4a, p_ba3a, p_tle_standards, p_tle_blanks, ncol = 1, align = "hv", axis = "b")

p_tle_combined

cowplot::save_plot(
  filename = "fig_output/tle_combined_plot.pdf",
  plot = p_tle_combined,
  base_height = 12,
  base_width = 8
)

cowplot::save_plot(
  filename = "fig_output/tle_combined_plot.png",
  plot = p_tle_combined,
  base_height = 12,
  base_width = 8
)

cowplot::save_plot(
  filename = "fig_output/ba1b_tle.png", plot = p_ba1b,
  base_height = 4, base_width = 6
)

cowplot::save_plot(
  filename = "fig_output/ba4a_tle.png", plot = p_ba4a,
  base_height = 4, base_width = 6
)

cowplot::save_plot(
  filename = "fig_output/ba3a_tle.png", plot = p_ba3a,
  base_height = 4, base_width = 6
)

cowplot::save_plot(
  filename = "fig_output/blanks_tle.png", plot = p_tle_blanks,
  base_height = 4, base_width = 6
)

cowplot::save_plot(
  filename = "fig_output/stds_tle.png", plot = p_tle_standards,
  base_height = 4, base_width = 8
)
```

### Compare BA1B70 and black goo to TLE

```{r}
xanes_compare_TLE <- xanes_w_metadata |> 
  filter(
    str_detect(filename, "goo") |
      str_detect(filename, "JS_ink") |
      str_detect(filename, "PowderPillow2_TLE_xanes_set2_003_001_fl") | # BA1B_70_TLE
      str_detect(filename, "data/bl143_dec22/02_mu/BA1B_70_G_001_001_fl.mu") |
      str_detect(compound, "Glutathione")
  )

p_compare_TLE <-xanes_compare_TLE |> 
  mutate(
    offset = case_when(
      str_detect(filename, "black_goo") ~ 3,
      str_detect(filename, "BA1B_70_G") ~ 4,
      str_detect(filename, "JS_ink") ~ 2,
      str_detect(filename, "PowderPillow2_TLE_xanes_set2") ~ 5,
      str_detect(filename, "PowderPillow2_TLE_xanes_set3") ~ 1
    )
  ) |> 
  arrange(offset) |> 
  ggplot(
    aes(
      x = energy,
      y = samnorm + offset,
      color = filename
    )
  ) +
  geom_vline(xintercept = c(2472.7, 2474.2)) +
  geom_line(linewidth = 1) +
  ggsci::scale_color_npg() +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2490))+
  theme_xas()
p_compare_TLE

cowplot::save_plot(
  filename = "fig_output/TLE_compare_to_goo.png",
  plot = p_compare_TLE,
  base_height = 5,
  base_width = 9
)
```

What about drilling fluid?

```{r}
xanes_mud_dayglo <- xanes_w_metadata |> 
  filter(compound %in% c("Dayglo Tracer", "Drilling Mud"))

p_drilling <- xanes_mud_dayglo |> 
  group_by(compound) |> 
  mutate(offset = cur_group_id()) |> 
  ggplot(
    aes(
      x = energy,
      y = samnorm - 3*offset,
      color = compound
    )
  ) +
  geom_line(linewidth = 1) +
  geom_vline(xintercept = c(2481.3, 2482.5)) +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2500)) +
  labs(
    caption = "Vertical lines @ 2481.3, 2482.5"
  ) +
  theme_xas()
p_drilling

cowplot::save_plot(
  filename = "fig_output/drilling_xanes.png",
  plot = p_drilling,
  base_height = 5, base_width = 7
)
```

## Mineral Standards

```{r}
xanes_stds <- xanes_w_e0s |> 
  filter(!is_TLE, is_standard)

xanes_stds_e0s <- xanes_stds |> 
  select(filename, compound, e0, e0_abs) |> 
  distinct() |> 
  arrange(e0) |> 
  mutate(offset = row_number())
```

### Plot mineral standards

```{r}
p_mineral_standards <- xanes_stds |> 
  left_join(xanes_stds_e0s, by = join_by(filename, compound)) |> 
  mutate(compound = fct_reorder(as.factor(compound), offset)) |> 
  arrange(offset) |> 
  ggplot(
    aes(
      x = energy,
      y = samnorm + offset,
      color = compound,
      fill = compound
    )
  ) +
  geom_vline(xintercept = c(2470.4, 2472.2, 2482.5), color = "gray", linewidth = 0.25) +
  geom_line(linewidth = 0.8, show.legend = FALSE) +
  # geom_point(
  #   data = xanes_stds_e0s,
  #   aes(
  #     x = e0, y = e0_abs + offset
  #   ),
  #   size = 2, color = "black", shape = 21
  # ) +
  geom_text(
    data = function(df) df |> select(offset, compound) |> distinct(),
    aes(label = compound, y = offset + 1.5),
    x = 2495, color = "black", size = 3, hjust = 1
  ) +
  ggprism::annotation_ticks(sides = "tb") +
  scale_color_viridis_d(end = 0.85) +
  scale_fill_viridis_d(end = 0.85) +
  coord_cartesian(xlim = c(2465, 2495)) +
  labs(
    x = "Energy (eV)",
    y = "",
    color = "",
    caption = "Vertical lines @ 2470.4, 2472.2, 2482.5"
  ) +
  theme_bw() +
  theme(
    axis.ticks.y = element_blank(),
    panel.grid.major.y = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    axis.text.y = element_blank()
  ) +
  guides(color = guide_legend(reverse=TRUE),
         fill = guide_legend(reverse=TRUE))

p_mineral_standards

cowplot::save_plot(
  filename = "fig_output/mineral_standards.pdf",
  plot = p_mineral_standards,
  base_height = 14, base_width = 4
)
```

# Reaction products

```{r}
xanes_rxns <- xanes_w_metadata |> 
  filter(str_detect(compound, "Brucite"))
```

```{r}
# Plot
xanes_rxns |> 
  mutate(offset = case_when(
    str_detect(compound, "Sulfide") ~ 1,
    str_detect(compound, "Thio") ~ 2
  )) |> 
  ggplot(
    aes(x = energy, y = samnorm + offset, color = compound)
  ) +
  geom_line() +
  ggprism::annotation_ticks(sides = "tb") +
  theme_xas() +
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.title.y = element_blank()
  )

# compare to tochilinite
tochilinite_xas <- xanes_w_metadata |> filter(compound %in% c("Sodium thiosulfate", "Tochilinite"))

p_rxns <- xanes_rxns |> 
  bind_rows(tochilinite_xas) |> 
  mutate(offset = case_when(
    compound == "Brucite + Sulfide Product" ~ 2,
    compound == "Brucite + Thiosulfate Product" ~ 1,
    compound == "Sodium thiosulfate" ~ 0,
    compound == "Tochilinite" ~ 3
  )) |> 
  ggplot(
    aes(x = energy, y = samnorm + offset/2, color = compound)
  ) +
  geom_vline(xintercept = c(2470.4, 2472.2, 2482.6)) +
  geom_line(linewidth = 1) +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2500)) +
  theme_xas() +
  theme(
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.title.y = element_blank()
  )
p_rxns


cowplot::save_plot(
  filename = "fig_output/sulfide_reaction_products.png",
  plot = p_rxns,
  base_height = 6, base_width = 7
)
```

# Core powders

```{r}
xanes_core_powders <- xanes_w_e0s |> 
  filter(is_powder, !is_standard, !is.na(core)) |> 
  mutate(
    sulfidization_time = case_when(
      !is_sulfidized ~ "Pre-Sulfidization",
      is_sulfidized & str_detect(filename, "dec22") ~ "5 months",
      is_sulfidized & str_detect(filename, "dec23") ~ "17 months",
      is_sulfidized & str_detect(filename, "jun24") ~ "23 months",
    )
  ) |> 
  mutate(
    core_depth_str = paste0(core, "-", core_depth_m)
  )
```

## BA4A Sulfidization Products

```{r}
p_ba4a_splus <- xanes_core_powders  |> 
  filter(
    core == "BA4A", 
    core_depth_m %in% c(10, 150)
    ) |> 
  filter(!str_detect(filename, "G_pre")) |> 
  mutate(sample_id = paste0(core, "-", core_depth_m, "m :", sulfidization_time)) |> 
  ggplot(
    aes(
      x = energy, y = samnorm,
      group = filename, color = sample_id
    )
  ) +
  geom_line() +
  facet_wrap(vars(core_depth_str), ncol = 1, scales = "free_y") +
  coord_cartesian(xlim = c(2460, 2500)) +
  ggprism::annotation_ticks(sides = "tb") +
  labs(
    title = "BA4A Sulfidization"
  ) +
  theme_xas()
p_ba4a_splus

# cowplot::save_plot(filename = "fig_output/ba4a_splus.png",
#                    plot = p_ba4a_splus,
#                    base_width = 7, base_height = 5)

```

Focus on BA4A_10

```{r}
p_ba4a_10_splus <- xanes_core_powders  |> 
  filter(core == "BA4A", core_depth_m == 10) |> 
  mutate(sample_id = as.factor(paste0(core, "-", core_depth_m, "m :", sulfidization_time))) |> 
  mutate(
    offset = case_when(
      str_detect(sulfidization_time, "Pre") ~ 0,
      str_detect(sulfidization_time, "5 months") ~ 1,
      str_detect(sulfidization_time, "17 months") ~ 2,
      str_detect(sulfidization_time, "23 months") ~ 3,
    ),
    sample_id = fct_reorder(sample_id, offset)
  ) |> 
  ggplot(
    aes(
      x = energy, y = samnorm + offset,
      group = filename, color = sample_id
    )
  ) +
  geom_vline(xintercept = 2470.4, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2472.4, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2482.5, color = "gray", linewidth = 0.5) +
  geom_line(linewidth = 1) +
  scale_color_manual(values = c("#9cc2ff", "#6197ed", "#6585ba", "#021d47")) +
  geom_text(
    data = function(df) df |> group_by(filename) |> slice(1) |> ungroup(),
    aes(label = sample_id, x = 2520, y = samnorm + offset+1.3),
    size = 3, hjust = 1,
    color = "black"
  ) +
  facet_wrap(vars(core_depth_str), ncol = 1, scales = "free_y") +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2500)) +
  labs(
    x = "Energy (eV)", y = "",
    title = "BA4A 10m Sulfidization",
    caption = "Bulk XANES data \nVertical lines are marked at 2470.4, 2472.4, and 2482.5 eV, respectively."
  ) +
  theme_xas() +
  theme(legend.position = "None")
p_ba4a_10_splus

# cowplot::save_plot(
#   filename = "fig_output/ba4a_10_splus.png",
#   plot = p_ba4a_10_splus, base_height = 5, base_width = 7
# )

```

Focus on BA4A 150

```{r}
p_ba4a_150_splus <- xanes_core_powders |> 
  filter(core == "BA4A", core_depth_m == 150) |> 
  mutate(sample_id = paste0(core, "-", core_depth_m, "m :", sulfidization_time)) |> 
  filter(!str_detect(filename, "G_pre")) |> 
  mutate(
    offset = case_when(
      str_detect(sulfidization_time, "Pre") ~ 0,
      str_detect(sulfidization_time, "5 months") ~ 1,
      str_detect(sulfidization_time, "17 months") ~ 2,
    ),
    sample_id = fct_reorder(sample_id, offset)
  ) |> 
  ggplot(
    aes(
      x = energy, y = samnorm + offset,
      group = filename, color = sample_id
    )
  ) +
  geom_vline(xintercept = 2470.4, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2472.4, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2482.5, color = "gray", linewidth = 0.5) +
  geom_line() +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2520)) +
  labs(
    title = "BA4A 150 Sulfidization"
  ) +
  theme_xas()
p_ba4a_150_splus

# cowplot::save_plot(
#   filename = "fig_output/ba4a_150_splus.png",
#   plot = p_ba4a_150_splus, base_height = 5, base_width = 7
# )
```

## BA1B Sulfidization Products

BA1B all samples

```{r}
p_ba1b_splus <- xanes_core_powders |> 
  filter(core == "BA1B", core_depth_m %in% c(20, 110, 140)) |> 
  mutate(sample_id = paste0(core, "-", core_depth_m, "m :", sulfidization_time)) |> 
  
  ggplot(
    aes(
      x = energy, y = samnorm,
      group = filename, color = sample_id
    )
  ) +
  geom_vline(xintercept = 2470.4, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2472.5, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2482.5, color = "gray", linewidth = 0.5) +
  geom_line() +
  facet_wrap(vars(is_sulfidized), ncol = 1) +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2520)) +
  labs(
    title = "BA1B Sulfidization"
  ) +
  theme_xas()
p_ba1b_splus

# cowplot::save_plot(
#   filename = "fig_output/ba1b_splus.png",
#   plot = p_ba1b_splus,
#   base_height = 5, base_width = 7
# )
```

Focus on BA1B 20

```{r}
xanes_core_powders |> 
  filter(core == "BA1B", core_depth_m == 20) |> 
  mutate(sample_id = paste0(core, "-", core_depth_m, "m :", sulfidization_time)) |> 
  mutate(
    offset = case_when(
      str_detect(sulfidization_time, "Pre") ~ 0,
      str_detect(sulfidization_time, "23 months") ~ 2,
    ),
    sample_id = fct_reorder(sample_id, offset)
  ) |> 
  ggplot(
    aes(
      x = energy, y = samnorm + offset,
      group = filename, color = sample_id
    )
  ) +
  geom_vline(xintercept = 2470.4, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2472.5, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2482.5, color = "gray", linewidth = 0.5) +
  geom_line() +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2520)) +
  labs(
    title = "BA1B 20 Sulfidization"
  ) +
  theme_xas()
```

Focus on BA1B 110

```{r}
p_ba1b_110 <- xanes_core_powders |> 
  filter(core == "BA1B", core_depth_m == 110) |> 
  mutate(sample_id = paste0(core, "-", core_depth_m, "m :", sulfidization_time)) |> 
  mutate(
    offset = case_when(
      str_detect(sulfidization_time, "Pre") ~ 0,
      str_detect(sulfidization_time, "23 months") ~ 2,
    ),
    sample_id = fct_reorder(sample_id, offset)
  ) |> 
  ggplot(
    aes(
      x = energy, y = samnorm + offset,
      group = filename, color = sample_id
    )
  ) +
  geom_vline(xintercept = 2470.4, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2472.5, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2482.5, color = "gray", linewidth = 0.5) +
  geom_line() +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2520)) +
  labs(
    title = "BA1B 110 Sulfidization"
  ) +
  theme_xas()

p_ba1b_110

# cowplot::save_plot(
#   filename = "fig_output/ba1b_110_splus.png",
#   plot = p_ba1b_110, base_height = 5, base_width = 7
# )
```

Focus on BA1B 140

```{r}
p_ba1b_140 <- xanes_core_powders |> 
  filter(core == "BA1B", core_depth_m == 140) |> 
  mutate(sample_id = paste0(core, "-", core_depth_m, "m :", sulfidization_time)) |> 
  mutate(
    offset = case_when(
      str_detect(sulfidization_time, "Pre") ~ 0,
      str_detect(sulfidization_time, "23 months") ~ 2,
    ),
    sample_id = fct_reorder(sample_id, offset)
  ) |> 
  ggplot(
    aes(
      x = energy, y = samnorm + offset,
      group = filename, color = sample_id
    )
  ) +
  geom_vline(xintercept = 2470.4, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2472.5, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2482.5, color = "gray", linewidth = 0.5) +
  geom_line() +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2520)) +
  labs(
    title = "BA1B 140 Sulfidization"
  ) +
  theme_xas()
p_ba1b_140

# cowplot::save_plot(
#   filename = "fig_output/p_ba1b_140_splus.png",
#   plot = p_ba1b_140,
#   base_height = 5, base_width = 7
# )
```

## BA3A sulfidization products

```{r}
p_ba3a_splus <- xanes_core_powders |> 
  filter(core == "BA3A") |> 
  mutate(sample_id = paste0(core, "-", core_depth_m, "m :", sulfidization_time)) |> 
  ggplot(
    aes(
      x = energy, y = samnorm,
      group = filename, color = sample_id
    )
  ) +
  geom_vline(xintercept = 2470.4, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2472.5, color = "gray", linewidth = 0.5) +
  geom_vline(xintercept = 2482.5, color = "gray", linewidth = 0.5) +
  geom_line() +
  facet_wrap(vars(is_sulfidized), ncol = 1) +
  ggprism::annotation_ticks(sides = "tb") +
  coord_cartesian(xlim = c(2460, 2520), ylim = c(-3, 3)) +
  labs(
    title = "BA3A Sulfidization"
  ) +
  theme_xas()
p_ba3a_splus

# cowplot::save_plot(
#   filename = "fig_output/ba3a_splus.png",
#   plot = p_ba3a_splus, 
#   base_height = 5, base_width = 7
# )
```


Sulfidization data appears inconclusive at the bulk scale. The only sample of note is BA4A - 10m, which appears to show a reaction progression from oxidized sulfur to reduced sulfur, then to some kind of "disulfide"-like phase at approx. 2472 eV.

The remaining BA1B and BA4A samples have too much sulfur in them to show any change, whereas the BA3A samples have too little sulfur such that the SNR is poor.

Perhaps we can use the relative amplitudes of the monoS and DiS peaks to assess degree of sulfidization

```{r}
xanes_core_powders |> pull(energy) |> unique()

p_ba1b_140
p_ba1b_splus

plotly::ggplotly()
```

## Sulfidization comparison

Export all the rock core powders to baseline subtract them in origin

```{r}
# xanes_core_powders |> 
#   select(energy, samnorm, filename, sample_id) |> 
#   group_by(sample_id) |> 
#   group_walk(~writexl::write_xlsx(.x, paste0("data/spectra_for_origin/", .y$sample_id, ".xlsx")))

```

Import all the baseline-subtracted XANES from origin
```{r}
cnames <- readxl::read_excel("data/spectra_for_origin/BA1B_110_G_006_001_SUBBED.xlsx", n_max = 0) |> 
    names()

read_many_xlsx <- function(directory) {
  directory |> 
    map(readxl::read_xlsx, skip = 2, col_names = cnames) |> 
    bind_rows(.id = "file")
}

# read in all the files
blsub.fp <- "data/origin_baseline_subtracted/"
blsub.files <- list.files(path = blsub.fp, full.names = TRUE)

xanes_blsub <- read_many_xlsx(directory = blsub.files) |> 
  rename(
    samnorm_blsub = `Subtracted_Data Y1`,
    samnorm_baseline = `Baseline_Data Y1`
    ) |> 
  select(-c(`Subtracted_Data X1`,`Baseline_Data X1`, file, samnorm)) |> 
  left_join(
    xanes_core_powders, by = join_by(filename, energy)
  )


```

Find appropriate peak windows
```{r}
xanes_blsub |> 
  filter(
    core == "BA4A", 
    core_depth_m %in% c(10, 150)
    ) |> 
  filter(!str_detect(filename, "G_pre")) |> 
  mutate(sample_id = paste0(core, "-", core_depth_m, "m :", sulfidization_time)) |> 
  ggplot(
    aes(
      x = energy, y = samnorm_blsub,
      group = filename, color = sample_id
    )
  ) +
  geom_line() +
  geom_vline(xintercept = c(2469, 2473)) +
  geom_vline(xintercept = c(2481, 2484)) +
  facet_wrap(vars(core_depth_str), ncol = 1, scales = "free_y") +
  coord_cartesian(xlim = c(2460, 2500)) +
  ggprism::annotation_ticks(sides = "tb") +
  labs(
    title = "BA4A Sulfidization"
  ) +
  theme_xas()
```

```{r}
sulfide_to_sulfate_ratio <- xanes_blsub |> 
  group_by(filename) |> 
  summarise(
    monoS_area = sum(ifelse(energy > 2469 & energy < 2473, samnorm_blsub, NA), na.rm = TRUE),
    so4_area = sum(ifelse(energy > 2481 & energy < 2484, samnorm_blsub, NA), na.rm = TRUE)
  ) |> 
  mutate(
    sulfide_to_sulfate_ratio = monoS_area / so4_area
  ) |> 
  left_join(metadata, by = join_by(filename))

sulfide_to_sulfate_ratio |> 
  mutate(
    sulfidization_time = case_when(
      !is_sulfidized ~ "Pre-Sulfidization",
      is_sulfidized & str_detect(filename, "dec22") ~ "5 months",
      is_sulfidized & str_detect(filename, "dec23") ~ "17 months",
      is_sulfidized & str_detect(filename, "jun24") ~ "23 months",
    ),
    sulfidization_time = factor(
      sulfidization_time, levels = c("Pre-Sulfidization", "5 months", "17 months","23 months"))
  ) |> 
  mutate(
    core_depth_str = paste0(core, "-", core_depth_m)
  ) |> 
  ggplot(
    aes(
      x = sulfidization_time,
      y = sulfide_to_sulfate_ratio,
      color = core_depth_str,
      group = core_depth_str
    )
  ) +
  geom_point() +
  geom_line() +
  facet_wrap(vars(core), ncol = 1, scales = "free")
```
